[NOX-52] create library for examples' common functionality
