+++ examples/CMakeLists.txt
@@ -1,5 +1,6 @@
 set(NOX_EXAMPLES_THIRD_PARTY_FOLDER_NAME "${NOX_EXAMPLES_FOLDER_NAME}/third_party")
 set(NOX_EXAMPLES_THIRD_PARTY_DIR "${PROJECT_SOURCE_DIR}/examples/third_party")
+set(NOX_EXAMPLES_BASE_LIB_NAME "base")
 
 macro(add_example_project NAME)
     # --------------------------------------------------
@@ -20,26 +21,12 @@ macro(add_example_project NAME)
     # --------------------------------------------------
     # Link libraries
     # --------------------------------------------------
-    set(GLFW_BUILD_DOCS OFF CACHE BOOL "" FORCE)
-    set(GLFW_INSTALL OFF CACHE BOOL "" FORCE)
-    add_subdirectory(${NOX_EXAMPLES_THIRD_PARTY_DIR}/glfw glfw)
-    set_target_properties(
-        glfw
-        PROPERTIES 
-        FOLDER "${NOX_EXAMPLES_THIRD_PARTY_FOLDER_NAME}/glfw"
-    )
-    set_target_properties(
-        update_mappings
-        PROPERTIES
-        FOLDER "${NOX_EXAMPLES_THIRD_PARTY_FOLDER_NAME}/glfw"
-    )
-
     target_link_libraries(
         ${NAME}
         PRIVATE
-        ${NOX_LIB_NAME}
-        glfw
+        ${NOX_EXAMPLES_BASE_LIB_NAME}
     )
 endmacro()
 
+add_subdirectory(base)
 add_subdirectory(sandbox)
