+++ examples/base/src/CMakeLists.txt
@@ -1,0 +1,7 @@
+set(NOX_EXAMPLE_BASE_SOURCES
+    ${CMAKE_CURRENT_SOURCE_DIR}/CMakeLists.txt
+    ${CMAKE_CURRENT_SOURCE_DIR}/example.cpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/window.cpp
+)
+
+target_sources(${NOX_EXAMPLES_BASE_LIB_NAME} PRIVATE ${NOX_EXAMPLE_BASE_SOURCES})
