+++ src/opengl/linux/x11_gl_context.cpp
@@ -12,7 +12,7 @@ X11GLContext::X11GLContext(const X11SurfaceBackendDescriptor &descriptor) {
 }
 
 X11GLContext::~X11GLContext() {
-    [[maybe_unused]] XWindowAttributes windowAttributes; //NOLINT
+    [[maybe_unused]] XWindowAttributes windowAttributes; // NOLINT
     NOX_ASSERT_MSG(XGetWindowAttributes(m_handleDisplayX11, m_handleWindowX11, &windowAttributes),
                    "The window handle associated with OpenGL context is invalid");
